move timedOutQueryId handling inside insightDataLogic
