+++ frontend/src/scenes/insights/insightDataLogic.ts
@@ -1,3 +1,4 @@
+import posthog from 'posthog-js'
 import { actions, connect, kea, key, listeners, path, props, propsChanged, reducers, selectors } from 'kea'
 import { FilterType, InsightLogicProps, InsightType } from '~/types'
 import { keyForInsightLogicProps } from 'scenes/insights/sharedUtils'
@@ -18,6 +19,9 @@ import { compareFilters } from './utils/compareFilters'
 import { filterTestAccountsDefaultsLogic } from 'scenes/project/Settings/filterTestAccountDefaultsLogic'
 import { insightDataTimingLogic } from './insightDataTimingLogic'
 import { teamLogic } from 'scenes/teamLogic'
+import { sceneLogic } from 'scenes/sceneLogic'
+
+const SHOW_TIMEOUT_MESSAGE_AFTER = 5000
 
 const queryFromFilters = (filters: Partial<FilterType>): InsightVizNode => ({
     kind: NodeKind.InsightVizNode,
@@ -62,6 +66,7 @@ export const insightDataLogic = kea<insightDataLogicType>([
 
     actions({
         setQuery: (query: Node | null) => ({ query }),
+        setTimedOutQueryId: (id: string | null) => ({ id }),
         saveInsight: (redirectToViewMode = true) => ({ redirectToViewMode }),
         toggleQueryEditorPanel: true,
         cancelChanges: true,
@@ -78,6 +83,12 @@ export const insightDataLogic = kea<insightDataLogicType>([
             false,
             {
                 toggleQueryEditorPanel: (state) => !state,
+            },
+        ],
+        timedOutQueryId: [
+            null as null | string,
+            {
+                setTimedOutQueryId: (_, { id }) => id,
             },
         ],
     }),
@@ -207,6 +218,26 @@ export const insightDataLogic = kea<insightDataLogicType>([
             const savedResult = values.savedInsight.result
             actions.setQuery(savedFilters ? queryFromFilters(savedFilters) : null)
             actions.setInsightData({ ...values.insightData, result: savedResult ? savedResult : null })
+        },
+        loadData: async ({ queryId }, breakpoint) => {
+            actions.setTimedOutQueryId(null)
+
+            await breakpoint(SHOW_TIMEOUT_MESSAGE_AFTER)
+
+            if (values.insightDataLoading) {
+                actions.setTimedOutQueryId(queryId)
+                const tags = {
+                    kind: values.query?.kind,
+                    scene: sceneLogic.isMounted() ? sceneLogic.values.scene : null,
+                }
+                posthog.capture('insight timeout message shown', tags)
+            }
+        },
+        loadDataSuccess: () => {
+            actions.setTimedOutQueryId(null)
+        },
+        loadDataFailure: () => {
+            actions.setTimedOutQueryId(null)
         },
     })),
     propsChanged(({ actions, props, values }) => {
