+++ frontend/src/scenes/trends/trendsDataLogic.ts
@@ -7,6 +7,7 @@ import type { trendsDataLogicType } from './trendsDataLogicType'
 import { IndexedTrendResult } from './types'
 import { insightVizDataLogic } from 'scenes/insights/insightVizDataLogic'
 import { dayjs } from 'lib/dayjs'
+import { insightDataLogic } from 'scenes/insights/insightDataLogic'
 
 export const trendsDataLogic = kea<trendsDataLogicType>([
     props({} as InsightLogicProps),
@@ -38,7 +39,7 @@ export const trendsDataLogic = kea<trendsDataLogicType>([
                 'hasLegend',
             ],
         ],
-        actions: [insightVizDataLogic(props), ['setInsightData', 'updateInsightFilter']],
+        actions: [insightDataLogic(props), ['setInsightData'], insightVizDataLogic(props), ['updateInsightFilter']],
     })),
 
     actions({
